Fixed Certificate upgrade problems
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@2111 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/se/anatom/ejbca/ca/store/certificateprofiles/CertificateProfile.java
@@ -28,12 +28,12 @@ import se.anatom.ejbca.util.UpgradeableDataHashMap;
  * CertificateProfile is a basic class used to customize a certificate
  * configuration or be inherited by fixed certificate profiles.
  *
- * @version $Id: CertificateProfile.java,v 1.24 2004-11-16 21:55:16 herrvendil Exp $
+ * @version $Id: CertificateProfile.java,v 1.25 2004-11-20 16:59:59 herrvendil Exp $
  */
 public class CertificateProfile extends UpgradeableDataHashMap implements Serializable, Cloneable {
     private static Logger log = Logger.getLogger(CertificateProfile.class);
     // Default Values
-    public static final float LATEST_VERSION = 9;
+    public static final float LATEST_VERSION = (float) 9.0;
 
     /** KeyUsage constants */
     public static final int DIGITALSIGNATURE = 0;
